code reviews

+++ geov2/query_contains.go
@@ -92,7 +92,7 @@ func (cq *containsQuery) Evaluate(geoData segment.GeoShapeV2Data) *util.Bitset {
 	// filter out any maybeHits that do not have a bounding box that
 	// contains the query bounding box
 	boxFilter := func(docNum int) {
-		docBBoxBytes, err := geoData.BoundingBox(uint64(docNum))
+		docBBoxBytes, err := geoData.BoundingBox(uint32(docNum))
 		if docBBoxBytes == nil || err != nil {
 			return
 		}
@@ -112,7 +112,7 @@ func (cq *containsQuery) Evaluate(geoData segment.GeoShapeV2Data) *util.Bitset {
 	// filter out any maybeHits that do not have a shape that
 	// contains the query shape
 	shapeFilter := func(docNum int) {
-		docShapeBytes, err := geoData.Shape(uint64(docNum))
+		docShapeBytes, err := geoData.Shape(uint32(docNum))
 		if docShapeBytes == nil || err != nil {
 			return
 		}
