+++ geov2/query_intersects.go
@@ -96,7 +96,7 @@ func (iq *intersectsQuery) Evaluate(geoData segment.GeoShapeV2Data) *util.Bitset
 	// filter out any maybeHits that do not have a bounding box that
 	// intersects the query bounding box
 	boxFilter := func(docNum int) {
-		docBBoxBytes, err := geoData.BoundingBox(uint64(docNum))
+		docBBoxBytes, err := geoData.BoundingBox(uint32(docNum))
 		if docBBoxBytes == nil || err != nil {
 			return
 		}
@@ -116,7 +116,7 @@ func (iq *intersectsQuery) Evaluate(geoData segment.GeoShapeV2Data) *util.Bitset
 	// filter out any maybeHits that do not have a shape that
 	// intersects the query shape
 	shapeFilter := func(docNum int) {
-		docShapeBytes, err := geoData.Shape(uint64(docNum))
+		docShapeBytes, err := geoData.Shape(uint32(docNum))
 		if docShapeBytes == nil || err != nil {
 			return
 		}
